Change the popover frame so that the arrow shows scroll view
By extending the popover window to the full height of the "popover
frame" including the arrow, we can then mask the header to look like the
popover is the same shape as before, but so the visual effect view
header actually draws the arrow as well. We use a static image as a
mask to make the top roundrect shape with the arrow. Perhaps it would be
better to draw this in code, but it likely won't change shape very
often. This required some changes to INPopoverWindow since only changing
the arrow size wouldn't quite give us what we want.