Fix non-clipping scroll view sticking past the top of the popover
This was actually an easy fix that avoids the problem by just moving the top of the
scrollview down! Because of the way the header visual effect view works, it
still blurs the content underneath through the entire height of itself, but
avoids the possibility of the scroll view sticking past. This required the
change of the table view content insets as well.