Large refactor to extract recording management from VC
Recordings are now managed by a JEFRecordingManager that gets passed into the
content VC and recording table view VC. The header that overlays the table view
is now also part of the parent content VC in preparation for moving the
preferences view into the window as well (and it would transition like a
UINavigationController). Only rudimentary testing has occured at this commit,
but I needed a save point.