Fix differing heights between recording and other views
For some reason the recordings view was taller, even though they were all set to
the same height. This is the first way that works (adding bottom padding to the
scroll view) but I'm not sure *why* it fixes the issue. There's a masking
artifact in the bottom right corner from the scroll view when transitioning VCs,
but this is better than nothing.